Use ColumnHandle interface in projection adaptation classes
