+++ presto-hive/src/main/java/io/prestosql/plugin/hive/HivePageSourceProvider.java
@@ -13,6 +13,9 @@
  */
 package io.prestosql.plugin.hive;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.BiMap;
+import com.google.common.collect.ImmutableBiMap;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import io.prestosql.orc.metadata.ColumnMetadata;
@@ -44,10 +47,12 @@ import org.apache.hadoop.fs.Path;
 
 import javax.inject.Inject;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.OptionalInt;
 import java.util.Properties;
@@ -260,7 +265,7 @@ public class HivePageSourceProvider
                 Optional<ReaderColumns> readerProjections = readerWithProjections.get().getReaderColumns();
                 Optional<ReaderProjectionsAdapter> adapter = Optional.empty();
                 if (readerProjections.isPresent()) {
-                    adapter = Optional.of(new ReaderProjectionsAdapter(desiredColumns, readerProjections.get()));
+                    adapter = Optional.of(hiveProjectionsAdapter(desiredColumns, readerProjections.get()));
                 }
 
                 return Optional.of(new HivePageSource(
@@ -295,7 +300,7 @@ public class HivePageSourceProvider
                 Optional<ReaderColumns> projections = readerWithProjections.get().getProjectedReaderColumns();
 
                 if (projections.isPresent()) {
-                    ReaderProjectionsAdapter projectionsAdapter = new ReaderProjectionsAdapter(desiredColumns, projections.get());
+                    ReaderProjectionsAdapter projectionsAdapter = hiveProjectionsAdapter(desiredColumns, projections.get());
                     delegate = new HiveReaderProjectionsAdaptingRecordCursor(delegate, projectionsAdapter);
                 }
 
@@ -337,6 +342,37 @@ public class HivePageSourceProvider
         }
         Optional<HiveBucketFilter> hiveBucketFilter = getHiveBucketFilter(hiveTable, dynamicFilter.getCurrentPredicate());
         return hiveBucketFilter.map(filter -> !filter.getBucketsToKeep().contains(hiveSplit.getBucketNumber().getAsInt())).orElse(false);
+    }
+
+    private static ReaderProjectionsAdapter hiveProjectionsAdapter(List<HiveColumnHandle> expectedColumns, ReaderColumns readColumns)
+    {
+        return new ReaderProjectionsAdapter(
+                expectedColumns.stream().map(ColumnHandle.class::cast).collect(toImmutableList()),
+                readColumns,
+                column -> ((HiveColumnHandle) column).getType(),
+                HivePageSourceProvider::getProjection);
+    }
+
+    @VisibleForTesting
+    static List<Integer> getProjection(ColumnHandle expected, ColumnHandle read)
+    {
+        HiveColumnHandle expectedColumn = (HiveColumnHandle) expected;
+        HiveColumnHandle readColumn = (HiveColumnHandle) read;
+
+        checkArgument(expectedColumn.getBaseColumn().equals(readColumn.getBaseColumn()), "reader column is not valid for expected column");
+
+        List<Integer> expectedDereferences = expectedColumn.getHiveColumnProjectionInfo()
+                .map(HiveColumnProjectionInfo::getDereferenceIndices)
+                .orElse(ImmutableList.of());
+
+        List<Integer> readerDereferences = readColumn.getHiveColumnProjectionInfo()
+                .map(HiveColumnProjectionInfo::getDereferenceIndices)
+                .orElse(ImmutableList.of());
+
+        checkArgument(readerDereferences.size() <= expectedDereferences.size(), "Field returned by the reader should include expected field");
+        checkArgument(expectedDereferences.subList(0, readerDereferences.size()).equals(readerDereferences), "Field returned by the reader should be a prefix of expected field");
+
+        return expectedDereferences.subList(readerDereferences.size(), expectedDereferences.size());
     }
 
     public static class ColumnMapping
@@ -639,6 +675,152 @@ public class HivePageSourceProvider
         public int getBucketToKeep()
         {
             return bucketToKeep;
+        }
+    }
+
+    /**
+     * Creates a mapping between the input {@param columns} and base columns if required.
+     */
+    public static Optional<ReaderColumns> projectBaseColumns(List<HiveColumnHandle> columns)
+    {
+        requireNonNull(columns, "columns is null");
+
+        // No projection is required if all columns are base columns
+        if (columns.stream().allMatch(HiveColumnHandle::isBaseColumn)) {
+            return Optional.empty();
+        }
+
+        ImmutableList.Builder<ColumnHandle> projectedColumns = ImmutableList.builder();
+        ImmutableList.Builder<Integer> outputColumnMapping = ImmutableList.builder();
+        Map<Integer, Integer> mappedHiveColumnIndices = new HashMap<>();
+        int projectedColumnCount = 0;
+
+        for (HiveColumnHandle column : columns) {
+            int hiveColumnIndex = column.getBaseHiveColumnIndex();
+            Integer mapped = mappedHiveColumnIndices.get(hiveColumnIndex);
+
+            if (mapped == null) {
+                projectedColumns.add(column.getBaseColumn());
+                mappedHiveColumnIndices.put(hiveColumnIndex, projectedColumnCount);
+                outputColumnMapping.add(projectedColumnCount);
+                projectedColumnCount++;
+            }
+            else {
+                outputColumnMapping.add(mapped);
+            }
+        }
+
+        return Optional.of(new ReaderColumns(projectedColumns.build(), outputColumnMapping.build()));
+    }
+
+    /**
+     * Creates a set of sufficient columns for the input projected columns and prepares a mapping between the two. For example,
+     * if input {@param columns} include columns "a.b" and "a.b.c", then they will be projected from a single column "a.b".
+     */
+    public static Optional<ReaderColumns> projectSufficientColumns(List<HiveColumnHandle> columns)
+    {
+        requireNonNull(columns, "columns is null");
+
+        if (columns.stream().allMatch(HiveColumnHandle::isBaseColumn)) {
+            return Optional.empty();
+        }
+
+        ImmutableBiMap.Builder<DereferenceChain, HiveColumnHandle> dereferenceChainsBuilder = ImmutableBiMap.builder();
+
+        for (HiveColumnHandle column : columns) {
+            List<Integer> indices = column.getHiveColumnProjectionInfo()
+                    .map(HiveColumnProjectionInfo::getDereferenceIndices)
+                    .orElse(ImmutableList.of());
+
+            DereferenceChain dereferenceChain = new DereferenceChain(column.getBaseColumnName(), indices);
+            dereferenceChainsBuilder.put(dereferenceChain, column);
+        }
+
+        BiMap<DereferenceChain, HiveColumnHandle> dereferenceChains = dereferenceChainsBuilder.build();
+
+        List<ColumnHandle> sufficientColumns = new ArrayList<>();
+        ImmutableList.Builder<Integer> outputColumnMapping = ImmutableList.builder();
+
+        Map<DereferenceChain, Integer> pickedColumns = new HashMap<>();
+
+        // Pick a covering column for every column
+        for (HiveColumnHandle columnHandle : columns) {
+            DereferenceChain column = dereferenceChains.inverse().get(columnHandle);
+            List<DereferenceChain> orderedPrefixes = column.getOrderedPrefixes();
+            DereferenceChain chosenColumn = null;
+
+            // Shortest existing prefix is chosen as the input.
+            for (DereferenceChain prefix : orderedPrefixes) {
+                if (dereferenceChains.containsKey(prefix)) {
+                    chosenColumn = prefix;
+                    break;
+                }
+            }
+
+            checkState(chosenColumn != null, "chosenColumn is null");
+            int inputBlockIndex;
+
+            if (pickedColumns.containsKey(chosenColumn)) {
+                // Use already picked column
+                inputBlockIndex = pickedColumns.get(chosenColumn);
+            }
+            else {
+                // Add a new column for the reader
+                sufficientColumns.add(dereferenceChains.get(chosenColumn));
+                pickedColumns.put(chosenColumn, sufficientColumns.size() - 1);
+                inputBlockIndex = sufficientColumns.size() - 1;
+            }
+
+            outputColumnMapping.add(inputBlockIndex);
+        }
+
+        return Optional.of(new ReaderColumns(sufficientColumns, outputColumnMapping.build()));
+    }
+
+    private static class DereferenceChain
+    {
+        private final String name;
+        private final List<Integer> indices;
+
+        public DereferenceChain(String name, List<Integer> indices)
+        {
+            this.name = requireNonNull(name, "name is null");
+            this.indices = ImmutableList.copyOf(requireNonNull(indices, "indices is null"));
+        }
+
+        @Override
+        public boolean equals(Object o)
+        {
+            if (this == o) {
+                return true;
+            }
+            if (o == null || getClass() != o.getClass()) {
+                return false;
+            }
+
+            DereferenceChain that = (DereferenceChain) o;
+            return Objects.equals(name, that.name) &&
+                    Objects.equals(indices, that.indices);
+        }
+
+        @Override
+        public int hashCode()
+        {
+            return Objects.hash(name, indices);
+        }
+
+        /**
+         * Get Prefixes of this Dereference chain in increasing order of lengths
+         */
+        public List<DereferenceChain> getOrderedPrefixes()
+        {
+            ImmutableList.Builder<DereferenceChain> prefixes = ImmutableList.builder();
+
+            for (int prefixLen = 0; prefixLen <= indices.size(); prefixLen++) {
+                prefixes.add(new DereferenceChain(name, indices.subList(0, prefixLen)));
+            }
+
+            return prefixes.build();
         }
     }
 }
