+++ presto-hive/src/main/java/io/prestosql/plugin/hive/ReaderColumns.java
@@ -13,33 +13,29 @@
  */
 package io.prestosql.plugin.hive;
 
-import com.google.common.collect.BiMap;
-import com.google.common.collect.ImmutableBiMap;
 import com.google.common.collect.ImmutableList;
+import io.prestosql.spi.connector.ColumnHandle;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
 
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkState;
 import static java.util.Objects.requireNonNull;
 
 /**
- * Stores a mapping of the projected columns required by {@link HivePageSource} to the columns supplied by format-specific
- * page sources or record cursors.
+ * Stores a mapping between
+ *  - the projected columns required by a connector level pagesource and
+ *  -  the columns supplied by format-specific page source
+ *
+ * Currently used in {@link HivePageSource}.
  */
 public class ReaderColumns
 {
     // columns to be read by the reader (ordered)
-    private final List<HiveColumnHandle> readerColumns;
+    private final List<ColumnHandle> readerColumns;
     // indices for mapping expected hive column handles to the reader's column handles
     private final List<Integer> readerBlockIndices;
 
-    private ReaderColumns(List<HiveColumnHandle> readerColumns, List<Integer> readerBlockIndices)
+    public ReaderColumns(List<? extends ColumnHandle> readerColumns, List<Integer> readerBlockIndices)
     {
         this.readerColumns = ImmutableList.copyOf(requireNonNull(readerColumns, "readerColumns is null"));
 
@@ -48,9 +44,9 @@ public class ReaderColumns
     }
 
     /**
-     * For a column required by the {@link HivePageSource}, returns the column read by the delegate page source or record cursor.
+     * For a column required by the wrapper page source, returns the column read by the delegate page source or record cursor.
      */
-    public HiveColumnHandle getForColumnAt(int index)
+    public ColumnHandle getForColumnAt(int index)
     {
         checkArgument(index >= 0 && index < readerBlockIndices.size(), "index is not valid");
         int readerIndex = readerBlockIndices.get(index);
@@ -58,7 +54,7 @@ public class ReaderColumns
     }
 
     /**
-     * For a channel expected by {@link HivePageSource}, returns the channel index in the underlying page source or record cursor.
+     * For a channel expected by wrapper page source, returns the channel index in the underlying page source or record cursor.
      */
     public int getPositionForColumnAt(int index)
     {
@@ -69,154 +65,8 @@ public class ReaderColumns
     /**
      * returns the actual list of columns being read by underlying page source or record cursor in order.
      */
-    public List<HiveColumnHandle> get()
+    public List<ColumnHandle> get()
     {
         return readerColumns;
-    }
-
-    /**
-     * Creates a mapping between the input {@param columns} and base columns if required.
-     */
-    public static Optional<ReaderColumns> projectBaseColumns(List<HiveColumnHandle> columns)
-    {
-        requireNonNull(columns, "columns is null");
-
-        // No projection is required if all columns are base columns
-        if (columns.stream().allMatch(HiveColumnHandle::isBaseColumn)) {
-            return Optional.empty();
-        }
-
-        ImmutableList.Builder<HiveColumnHandle> projectedColumns = ImmutableList.builder();
-        ImmutableList.Builder<Integer> outputColumnMapping = ImmutableList.builder();
-        Map<Integer, Integer> mappedHiveColumnIndices = new HashMap<>();
-        int projectedColumnCount = 0;
-
-        for (HiveColumnHandle column : columns) {
-            int hiveColumnIndex = column.getBaseHiveColumnIndex();
-            Integer mapped = mappedHiveColumnIndices.get(hiveColumnIndex);
-
-            if (mapped == null) {
-                projectedColumns.add(column.getBaseColumn());
-                mappedHiveColumnIndices.put(hiveColumnIndex, projectedColumnCount);
-                outputColumnMapping.add(projectedColumnCount);
-                projectedColumnCount++;
-            }
-            else {
-                outputColumnMapping.add(mapped);
-            }
-        }
-
-        return Optional.of(new ReaderColumns(projectedColumns.build(), outputColumnMapping.build()));
-    }
-
-    /**
-     * Creates a set of sufficient columns for the input projected columns and prepares a mapping between the two. For example,
-     * if input {@param columns} include columns "a.b" and "a.b.c", then they will be projected from a single column "a.b".
-     */
-    public static Optional<ReaderColumns> projectSufficientColumns(List<HiveColumnHandle> columns)
-    {
-        requireNonNull(columns, "columns is null");
-
-        if (columns.stream().allMatch(HiveColumnHandle::isBaseColumn)) {
-            return Optional.empty();
-        }
-
-        ImmutableBiMap.Builder<DereferenceChain, HiveColumnHandle> dereferenceChainsBuilder = ImmutableBiMap.builder();
-
-        for (HiveColumnHandle column : columns) {
-            List<Integer> indices = column.getHiveColumnProjectionInfo()
-                    .map(HiveColumnProjectionInfo::getDereferenceIndices)
-                    .orElse(ImmutableList.of());
-
-            DereferenceChain dereferenceChain = new DereferenceChain(column.getBaseColumnName(), indices);
-            dereferenceChainsBuilder.put(dereferenceChain, column);
-        }
-
-        BiMap<DereferenceChain, HiveColumnHandle> dereferenceChains = dereferenceChainsBuilder.build();
-
-        List<HiveColumnHandle> sufficientColumns = new ArrayList<>();
-        ImmutableList.Builder<Integer> outputColumnMapping = ImmutableList.builder();
-
-        Map<DereferenceChain, Integer> pickedColumns = new HashMap<>();
-
-        // Pick a covering column for every column
-        for (HiveColumnHandle columnHandle : columns) {
-            DereferenceChain column = dereferenceChains.inverse().get(columnHandle);
-            List<DereferenceChain> orderedPrefixes = column.getOrderedPrefixes();
-            DereferenceChain chosenColumn = null;
-
-            // Shortest existing prefix is chosen as the input.
-            for (DereferenceChain prefix : orderedPrefixes) {
-                if (dereferenceChains.containsKey(prefix)) {
-                    chosenColumn = prefix;
-                    break;
-                }
-            }
-
-            checkState(chosenColumn != null, "chosenColumn is null");
-            int inputBlockIndex;
-
-            if (pickedColumns.containsKey(chosenColumn)) {
-                // Use already picked column
-                inputBlockIndex = pickedColumns.get(chosenColumn);
-            }
-            else {
-                // Add a new column for the reader
-                sufficientColumns.add(dereferenceChains.get(chosenColumn));
-                pickedColumns.put(chosenColumn, sufficientColumns.size() - 1);
-                inputBlockIndex = sufficientColumns.size() - 1;
-            }
-
-            outputColumnMapping.add(inputBlockIndex);
-        }
-
-        return Optional.of(new ReaderColumns(sufficientColumns, outputColumnMapping.build()));
-    }
-
-    private static class DereferenceChain
-    {
-        private final String name;
-        private final List<Integer> indices;
-
-        public DereferenceChain(String name, List<Integer> indices)
-        {
-            this.name = requireNonNull(name, "name is null");
-            this.indices = ImmutableList.copyOf(requireNonNull(indices, "indices is null"));
-        }
-
-        @Override
-        public boolean equals(Object o)
-        {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-
-            DereferenceChain that = (DereferenceChain) o;
-            return Objects.equals(name, that.name) &&
-                    Objects.equals(indices, that.indices);
-        }
-
-        @Override
-        public int hashCode()
-        {
-            return Objects.hash(name, indices);
-        }
-
-        /**
-         * Get Prefixes of this Dereference chain in increasing order of lengths
-         */
-        public List<DereferenceChain> getOrderedPrefixes()
-        {
-            ImmutableList.Builder<DereferenceChain> prefixes = ImmutableList.builder();
-
-            for (int prefixLen = 0; prefixLen <= indices.size(); prefixLen++) {
-                prefixes.add(new DereferenceChain(name, indices.subList(0, prefixLen)));
-            }
-
-            return prefixes.build();
-        }
     }
 }
